requirements.txt
asgiref==3.8.1
blinker==1.9.0
boto3==1.39.1
botocore==1.39.1
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.2.1
django-colorfield==0.14.0
django-storages==1.14.6
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.58.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
numpy<2.3.0  
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.9.0
zopfli==0.2.3.post1
python-dotenv==1.1.1
